refactor: first letter of category capitalized
make the first letter of each category camel case

<finishes>
