+++ src/__tests__/pages/layout-test.js
@@ -1,7 +1,7 @@
 import { shallow } from 'enzyme';
 import React from 'react';
 import toJson from 'enzyme-to-json';
-import Layout from '../../pages/Layout';
+import Layout from '../../pages/Layout.jsx';
 
 jest.mock('../../config/firebase', () => ({
 }));
