+++ src/pages/Layout.jsx
@@ -3,11 +3,11 @@ import { HashRouter as Router, Route, Switch } from 'react-router-dom';
 import React from 'react';
 import ReactLoading from 'react-loading';
 
-import Dashboard from './Dashboard';
+import Dashboard from './Dashboard.jsx';
 import firebaseAuth from '../config/firebase';
 import Footer from '../components/Footer.jsx';
 import Header from '../components/Header.jsx';
-import Login from './Login';
+import Login from './Login.jsx';
 import Page404 from './Page404.jsx';
 import { PrivateRoute } from '../components/PrivateRoute';
 import { PublicRoute } from '../components/PublicRoute';
