+++ src/style/mainStyle.scss
@@ -230,6 +230,7 @@ p {
 	margin-bottom: 1vw;
   margin-left: 2vw;
   width: 96vw;
+  text-transform: capitalize;
 }
 .sortByFeature {
 	width: 15vw;
